COM-2666: move profile admin sections to trainee follow up
