+++ src/core/components/courses/ProfileTraineeFollowUp.vue
@@ -2,6 +2,17 @@
   <div>
     <div class="q-mt-lg q-mb-xl">
       <p class="text-weight-bold">Émargements</p>
+      <div v-if="isIntraOrVendor" class="q-mb-xl">
+        <p class="text-weight-bold">Actions utiles</p>
+        <ni-banner v-if="followUpDisabled">
+          <template #message>
+            Il manque {{ formatQuantity('information', followUpMissingInfo.length ) }}
+            pour assurer le suivi de la formation : {{ followUpMissingInfo.join(', ') }}.
+          </template>
+        </ni-banner>
+        <ni-bi-color-button icon="file_download" label="Feuilles d'émargement"
+          :disable="disableDocDownload" @click="downloadAttendanceSheet" size="16px" />
+      </div>
       <attendance-table :course="course" />
     </div>
     <div v-if="areQuestionnaireAnswersVisible" class="q-mb-xl">
@@ -59,9 +70,14 @@ import QuestionnaireAnswersCell from '@components/courses/QuestionnaireAnswersCe
 import BiColorButton from '@components/BiColorButton';
 import Banner from '@components/Banner';
 import { SURVEY, OPEN_QUESTION, QUESTION_ANSWER, E_LEARNING } from '@data/constants';
-import { upperCaseFirstLetter, formatIdentity } from '@helpers/utils';
+import {
+  upperCaseFirstLetter,
+  formatIdentity,
+  formatQuantity,
+  readAPIResponseWithTypeArrayBuffer,
+} from '@helpers/utils';
 import { formatDate, ascendingSort, getTotalDuration, getDuration, formatIntervalHourly } from '@helpers/date';
-import { downloadZip } from '@helpers/file';
+import { downloadZip, downloadFile } from '@helpers/file';
 import { traineeFollowUpTableMixin } from '@mixins/traineeFollowUpTableMixin';
 import { courseMixin } from '@mixins/courseMixin';
 
@@ -118,6 +134,7 @@ export default {
   methods: {
     get,
     formatIntervalHourly,
+    formatQuantity,
     async refreshQuestionnaires () {
       try {
         this.questionnaires = await Courses.getCourseQuestionnaires(this.course._id);
@@ -175,6 +192,24 @@ export default {
       } catch (e) {
         console.error(e);
         NotifyNegative('Erreur lors du téléchargement des attestations.');
+      } finally {
+        this.pdfLoading = false;
+      }
+    },
+    async downloadAttendanceSheet () {
+      if (this.disableDocDownload) return;
+
+      try {
+        this.pdfLoading = true;
+        const pdf = await Courses.downloadAttendanceSheet(this.course._id);
+        downloadFile(pdf, 'emargement.pdf');
+      } catch (e) {
+        console.error(e);
+        if (e.status === 404) {
+          const { message } = readAPIResponseWithTypeArrayBuffer(e);
+          return NotifyNegative(message);
+        }
+        NotifyNegative('Erreur lors du téléchargement de la feuille d\'émargement.');
       } finally {
         this.pdfLoading = false;
       }
